+++ ChangeLog.md
@@ -12,6 +12,10 @@ Other enhancements:
 
 Bug fixes:
 
+* For versions of Cabal before 1.24, ensure that the dependencies of
+  non-buildable components are part of the build plan to work around an old
+  Cabal bug.
+
 ## v1.6.1
 
 Major changes: